requirements.txt
tensorflow==2.14.0
numpy
opencv-python
streamlit
pillow
